+++ src/gView.Framework.Core/Exceptions/MapServerException.cs
@@ -4,9 +4,17 @@ namespace gView.Framework.Core.Exceptions;
 
 public class MapServerException : Exception
 {
-    public MapServerException(string messsage)
-        : base(messsage)
+    public MapServerException(string message)
+        : base(message)
     {
-
+        
     }
+
+    public MapServerException(string message, int statusCode)
+        : this(message)
+    {
+        this.StatusCode = statusCode;
+    }
+
+    public int? StatusCode = null;
 }