+++ src/gView.Framework/Common/SystemInfo.cs
@@ -10,7 +10,7 @@ namespace gView.Framework.Common
 {
     public class SystemInfo
     {
-        public static Version Version = new Version(7, 25, 3301);
+        public static Version Version = new Version(7, 25, 3601);
 
         #region -> Private Variables