+++ src/gView.Server/AppCode/Extensions/ServiceMapExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using gView.Framework.Core.Carto;
+using gView.Framework.Core.Exceptions;
 using System;
 
 namespace gView.Server.AppCode.Extensions;
@@ -9,7 +10,17 @@ static internal class ServiceMapExtensions
     {
         if (serviceMap is null)
         {
-            throw new Exception("Can't load service. Check service logs for details");
+            throw new MapServerException("Can't load service. Check service logs for details");
+        }
+
+        return serviceMap;
+    }
+
+    static public IServiceMap ThrowIfNull(this IServiceMap serviceMap, string id)
+    {
+        if (serviceMap is null)
+        {
+            throw new MapServerException($"Unable to create map: {id}. Check log file for details", 500);
         }
 
         return serviceMap;